Cleaned-up and improved node code.
