+++ src/node/.arduino-ci.yml
@@ -8,7 +8,10 @@ platforms:
         - __AVR_ATmega328__
       warnings:
       flags:
-        - -coverage
+        - -g
+        - -O0
+        - --coverage
+        - -D__TEST__
 
 unittest:
   platforms:
